+++ example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_identity_kyc/flutter_identity_kyc.dart';
+import 'package:permission_handler/permission_handler.dart';
 
 void main() {
   runApp(MyApp());
@@ -14,7 +15,14 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     super.initState();
+    requestPermissions();
   }
+
+  Future<void> Function() requestPermissions = () async {
+    await Permission.camera.request();
+
+    await Permission.microphone.request();
+  };
 
   @override
   Widget build(BuildContext context) {
@@ -25,11 +33,11 @@ class _MyAppState extends State<MyApp> {
           ),
           body: Center(
             child: FlutterIdentityKyc(
-                merchantKey: "rvnn3ingodoed5ouy0z4:Cq_dXaVYS6F_R5HOQetA6HcLqKw",
+                merchantKey: "your public key",
                 email: "olayiwolakayode078@gmail.com",
                 firstName: "kayode",
                 lastName: "olayiwola",
-                isTest: true,
+                isTest: false,
                 userRef: "ddddd",
                 onCancel: (response) {
                   print(response);